add bokeh tests
